+++ ExchangeRatesService/Models/OpenExchangeRatesResult.cs
@@ -8,6 +8,6 @@ namespace ExchangeRatesService.Models
 		public string license { get; set; }
 		public int timestamp { get; set; }
 		public string @base { get; set; }
-		public Dictionary<string, decimal> rates { get; set; }
+		public Dictionary<string, double> rates { get; set; }
 	}
 }
